Update Forward Driver
This is the updated code. I have not made any method as static in the Forward class. The reason behind this being as I am taking the number of nodes, interpolating points,  the corresponding functional values and the point at which the functional value is required as user-inputs, I have to pass those arguments through some constructor for which I have to create an 'object'. Thus I see no smartness in calling those methods with the class name. Checked with various inputs. Working fine!