Create Simpson
This is a code written by Amit! The code can find the area for function like sin x,cos x, tan x, any polynomial, any polynomial involving the summation and subtraction of the above said functions. I have asked him to write a 'use me' file for the inputs to be given, and how by the use. This makes the error calculation a bit tough as the input is not known to the compiler. I tried to sort that even. Failed. If any one can, please solve it. Else, leave it. For it is a very good code.